Split group users component
